dpiスケールが整数になってて、まともに機能していなかった。

+++ src/HmGitWatcher/AsyncMessageBox.cs
@@ -60,7 +60,7 @@ internal partial class AsyncMessageBoxForm : Form
         }
 
         // 正しく取れたようだ
-        return dpi / 96;
+        return (double)dpi / (double)96;
     }
 
     // ----------------------------------------------------------------
